UOELSA-1303: Listataan kaikki saman erikoisalan kouluttajat yliopistosta riippumatta
